Cosmetic and Taskie Level implementation Interactable Refactoring

+++ demo/src/main/java/com/project/demo/logic/entity/LevelTaskie/LevelTaskieRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LevelTaskieRepository extends JpaRepository<TaskieLevel, Integer> {
+public interface LevelTaskieRepository extends JpaRepository<TaskieLevel, Long> {
     Optional<TaskieLevel> findByName(String levelName);
-    Optional<TaskieLevel> findById(Integer id);
+    Optional<TaskieLevel> findById(Long id);
 }
